disable SchoolRatings based on feature flag
